% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL.R
\name{load_transform_data}
\alias{load_transform_data}
\title{load_and_transform_data}
\usage{
load_transform_data()
}
\arguments{
\item{input}{}
}
\value{
this is main() function
}
\description{
load_and_transform_data
}
