% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL.R
\name{motherduck_df}
\alias{motherduck_df}
\title{motherduck_df}
\usage{
motherduck_df()
}
\arguments{
\item{input}{}
}
\value{
this function reads data from mother duck and pulls it in as a dataframe
}
\description{
motherduck_df
}
